requirements.txt
speedtest-cli
pandas